requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
idna==2.8
idna-ssl==1.1.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
multidict==4.5.2
mysqlclient==1.4.2.post1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
-e git+https://kuanpern@bitbucket.org/kuanpern/open_data_sciences.git@d48bc93993f7f41b481f3cac9a8a6684f56c83d3#egg=opends
openpyxl==2.6.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
scipy==1.3.0
Send2Trash==1.5.0
sendgrid==5.4.1
six==1.12.0
slackclient==2.1.0
SQLAlchemy==1.3.5
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
XlsxWriter==1.1.8
yarl==1.3.0
pymysql==0.9.3
python-dotenv==0.10.3
